+++ config/routes.rb
@@ -28,7 +28,7 @@ ActionController::Routing::Routes.draw do |map|
     page_mount.connect 'reset_password/*paths'
     page_mount.connect 'documents/*paths'
 
-    page_mount.connect 'booking-app-js/*paths'
+    page_mount.connect 'ecom-engine-js/*paths'
     page_mount.connect 'media/*paths'
   end
 end