+++ ecom_engine_extension.rb
@@ -1,9 +1,9 @@
 # Uncomment this if you reference any of your controllers in activate
 # require_dependency 'application_controller'
-class BookingAppExtension < Radiant::Extension
+class EcomEngineExtension < Radiant::Extension
   version "1.0"
-  description "This extension is communicating with Booking App"
-  url 'https://github.com/sealink/radiant-booking_app-extension'
+  description "This extension is communicating with Ecom Engine"
+  url 'https://github.com/sealink/radiant-ecom_engine-extension'
 
   def activate
   end